+++ git_deploy/cli.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python
 """
-git subcommand to remotely deploy projects. Manages merging of git branches.
+git subcommand to remotely deploy projects.
+
 Uses Ansible for the heavy lifting of deploying branches to remote servers,
 restarting appropriate services, and syncing static media to S3.
 """
@@ -16,6 +17,7 @@ WARNING = '\033[93m'
 FAIL = '\033[91m'
 ENDC = '\033[0m'
 
+
 ASSETS_DIR = os.environ.get('GIT_DEPLOY_ASSETS_DIR',
     os.path.expanduser('~/.git-deploy-assets'))
 
@@ -30,22 +32,9 @@ def sh(command, **kwargs):
     return _sh(*command.split())
 
 
-def get_playbooks_dir():
-    if os.environ.get('GIT_DEPLOY_PLAYBOOKS_DIR'):
-        return os.environ['GIT_DEPLOY_PLAYBOOKS_DIR']
-    else:
-        dirname = os.path.dirname(os.path.realpath(__file__))
-        return os.path.join(dirname, 'ansible')
-
-
 def get_vault_dir():
     return os.environ.get('GIT_DEPLOY_VAULT_DIR',
         os.path.join(ASSETS_DIR, 'vault'))
-
-
-def get_vault_password_file():
-    return os.environ.get('GIT_DEPLOY_VAULT_PASSWORD_FILE',
-        os.path.join(ASSETS_DIR, 'vault_password'))
 
 
 def get_project_path():
@@ -64,52 +53,25 @@ def call(command, **kwargs):
     return _call(command.split())
 
 
-#def usage():
-#    print('Usage:')
-#    envs = SUPPORTED_ENVIRONMENTS
-#    if len(envs.keys()) > 0:
-#        print('git-deploy (%s) [options]' % '|'.join(envs.keys()))
-#        print('\noptions:')
-#        print(' --verbose')
-#    else:
-#        print('No deployment environments found. ' \
-#            'Please add supported_envs to your Ansible configs')
-
-
 def get_common_config():
     fn = os.path.join(get_config_dir(), 'config.common.yml')
     with open(fn) as f:
         cfg = yaml.safe_load(f)
     return cfg
 COMMON_CONFIG = get_common_config()
-SUPPORTED_ENVIRONMENTS = COMMON_CONFIG['supported_envs']
 
 
 def ansible_playbook(env, playbook, *ansible_args, **kwargs):
-    # inventory is now based on ansible standards
-    # https://docs.ansible.com/ansible/latest/user_guide/intro_inventory.html
-    #hostfile = os.environ.get('GIT_DEPLOY_INVENTORY',
-    #    os.path.join(ASSETS_DIR, 'hosts'))
-    #playbooks_dir = get_playbooks_dir()
-    #if not playbook.startswith(os.sep):
-    #    playbook = os.path.join(playbooks_dir, playbook)
     command = 'ansible-playbook'
-    #if verbose:
-    #    command += ' -vvvv'
-    #if hostfile:
-    #    command += ' -i %s' % hostfile
     for a in ansible_args:
         command += f' {a}'
     command += ' -e env=%s' % env
-    command += ' -e project_root=%s' % get_project_path()
+    #command += ' -e project_root=%s' % get_project_path()
     command += ' -e config_dir=%s' % get_config_dir()
     command += ' -e vault_dir=%s' % get_vault_dir()
-    #command += ' -e playbooks_dir=%s' % playbooks_dir
     for k,v in kwargs.items():
         command += ' -e %s=%s' % (k,v)
     command += ' %s' % playbook
-    command += ' --vault-password-file %s' % get_vault_password_file()
-    print('COMMAND:', command)
     call(command)
 
 
@@ -120,23 +82,10 @@ def playbook_path(name):
 
 def builtin_playbook(env, playbook_name, *ansible_args):
     playbook = playbook_path(playbook_name)
-    ansible_playbook(env, playbook, *ansible_args, merge_from=SUPPORTED_ENVIRONMENTS[env])
+    ansible_playbook(env, playbook, *ansible_args)
 
 
-
-def deploy_extras(env, *ansible_args):
-    for fn in glob.glob(os.path.join(get_config_dir(), 'playbook*.yml')):
-        ansible_playbook(env, fn, *ansible_args)
-
-
-SUPPORTED_PLAYBOOKS = [
-    'deploy.python.yml',
-    'deploy.static.yml',
-    'deploy.repository.yml',
-    'deploy.web.yml'
-]
-
-def deploy(env, version, *ansible_args, project_virtualenv=None, playbook=None):
+def deploy(env, version, *ansible_args):
     gitdeploy_version = COMMON_CONFIG.get('gitdeploy_version')
     if gitdeploy_version and gitdeploy_version != __version__:
         print(FAIL + \
@@ -151,31 +100,16 @@ def deploy(env, version, *ansible_args, project_virtualenv=None, playbook=None):
         builtin_playbook(env, 'deploy.repository.yml', *ansible_args)
         builtin_playbook(env, 'deploy.static.yml', *ansible_args)
         return
-    if playbook is not None:
-        print('\nExecuting playbook: %s' % playbook)
-        if playbook not in SUPPORTED_PLAYBOOKS:
-            sys.exit("Unsupported playbook: %s. Options are: %s" % (
-                playbook, ', '.join(SUPPORTED_PLAYBOOKS)))
-        ansible_playbook(
-            env, playbook, *ansible_args, project_virtualenv=project_virtualenv)
+    builtin_playbook(env, 'deploy.repository.yml', *ansible_args)
+    if COMMON_CONFIG.get('docker_compose_file'):
+        builtin_playbook(env, 'build.containers.yml', *ansible_args)
+        builtin_playbook(env, 'deploy.web.yml', *ansible_args)
     else:
-        builtin_playbook(env, 'deploy.repository.yml', *ansible_args)
-        if COMMON_CONFIG.get('docker_compose_file'):
-            builtin_playbook(env, 'build.containers.yml', *ansible_args)
-            builtin_playbook(env, 'deploy.web.yml', *ansible_args)
-        else:
-            print('No containers to build. Deploying legacy application ...')
-            builtin_playbook(env, 'deploy.python.yml', *ansible_args)
-            builtin_playbook(env, 'deploy.static.yml', *ansible_args)
-            builtin_playbook(env, 'deploy.web.yml', *ansible_args)
+        print('No containers to build. Deploying legacy application ...')
+        builtin_playbook(env, 'deploy.python.yml', *ansible_args)
+        builtin_playbook(env, 'deploy.static.yml', *ansible_args)
+        builtin_playbook(env, 'deploy.web.yml', *ansible_args)
     print('\nDone')
-
-
-def main(env, version, *ansible_args):
-    project_virtualenv = os.environ.get('VIRTUAL_ENV')
-    playbook = None # TODO: support specific playbooks
-    deploy(env, version, *ansible_args, playbook=playbook,
-        project_virtualenv=project_virtualenv)
 
 
 import click
@@ -190,16 +124,21 @@ from click_option_group import optgroup, RequiredMutuallyExclusiveOptionGroup
 @click.version_option()
 @click.argument('env')
 @click.argument('project-version')
-@click.option('-x')
-def cli(ctx, env, project_version, x):
-    """Deploy to deployment environment ENV, PROJECT_VERSION of the current
-    project repository.
+@click.option('-p', '--playbook', required=False,
+    help='Specify a playbook to run rather than a project deployment.')
+def cli(ctx, env, project_version, playbook):
+    """Deploy to environment ENV, PROJECT_VERSION of the current project repository.
 
-    ENV is a configuration evironment with hosts configured in the ansible inventory.
+    ENV: A configuration evironment with hosts configured in the ansible inventory.
 
-    PROJECT_VERSION is either HEAD, or a branch or tag name.
+    PROJECT_VERSION: Either HEAD, or a branch or tag name.
+
+    ** Note: Additional arguments are passed (without validation) to Ansible.
     """
     ansible_args = ctx.args 
     ansible_args.extend(['-e', f'project_version={project_version}'])
     print('Passing arguments to ansible commands: %s\n' % ' '.join(ansible_args))
-    main(env, project_version, *ansible_args)
+    if playbook is not None:
+        ansible_playbook(env, playbook, *ansible_args)
+    else:
+        deploy(env, project_version, *ansible_args)